+++ mods/Farming/mod/prototypes/technology/tree_growing_technology.lua
@@ -27,8 +27,9 @@ create_technology({
     }
   },
   icon_size = 256,
-  extra_prerequisites = {"rs-oak-tree-patch"},
-  localized_name = {"recipe-name.rs-oak-growing"}
+  extra_prerequisites = {"rs-oak-tree-patch", "rs-oak-tree"},
+  localised_name = {"item-name.rs-oak-logs"},
+  localised_description = {"technology-description.rs-growing", {"tree.rs-oak-tree"}}
 })
 
 create_technology({
@@ -56,8 +57,9 @@ create_technology({
     }
   },
   icon_size = 256,
-  extra_prerequisites = {"rs-oak-growing", "rs-willow-tree-patch"},
-  localized_name = {"recipe-name.rs-willow-growing"}
+  extra_prerequisites = {"rs-oak-growing", "rs-willow-tree-patch", "rs-willow-tree"},
+  localised_name = {"item-name.rs-willow-logs"},
+  localised_description = {"technology-description.rs-growing", {"tree.rs-willow-tree"}}
 })
 
 create_technology({
@@ -85,8 +87,9 @@ create_technology({
     }
   },
   icon_size = 256,
-  extra_prerequisites = {"rs-willow-growing", "rs-maple-tree-patch"},
-  localized_name = {"recipe-name.rs-maple-growing"}
+  extra_prerequisites = {"rs-willow-growing", "rs-maple-tree-patch", "rs-maple-tree"},
+  localised_name = {"item-name.rs-maple-logs"},
+  localised_description = {"technology-description.rs-growing", {"tree.rs-maple-tree"}}
 })
 
 create_technology({
@@ -114,8 +117,9 @@ create_technology({
     }
   },
   icon_size = 256,
-  extra_prerequisites = {"rs-maple-growing", "rs-yew-tree-patch"},
-  localized_name = {"recipe-name.rs-yew-growing"}
+  extra_prerequisites = {"rs-maple-growing", "rs-yew-tree-patch", "rs-yew-tree"},
+  localised_name = {"item-name.rs-yew-logs"},
+  localised_description = {"technology-description.rs-growing", {"tree.rs-yew-tree"}}
 })
 
 create_technology({
@@ -143,8 +147,9 @@ create_technology({
     }
   },
   icon_size = 256,
-  extra_prerequisites = {"rs-yew-growing", "rs-magic-tree-patch"},
-  localized_name = {"recipe-name.rs-magic-growing"}
+  extra_prerequisites = {"rs-yew-growing", "rs-magic-tree-patch", "rs-magic-tree"},
+  localised_name = {"item-name.rs-magic-logs"},
+  localised_description = {"technology-description.rs-growing", {"tree.rs-magic-tree"}}
 })
 
 
@@ -175,7 +180,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-oak-growing"},
-  localized_name = {"recipe-name.rs-oak-growing-compost"}
+  localised_name = {"technology-name.rs-compost-growing", {"item-name.rs-oak-logs"}},
+  localised_description = {"technology-description.rs-compost-growing", {"tree.rs-oak-tree"}}
 })
 
 create_technology({
@@ -204,7 +210,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-oak-growing-compost", "rs-willow-growing"},
-  localized_name = {"recipe-name.rs-willow-growing-compost"}
+  localised_name = {"technology-name.rs-compost-growing", {"item-name.rs-willow-logs"}},
+  localised_description = {"technology-description.rs-compost-growing", {"tree.rs-willow-tree"}}
 })
 
 create_technology({
@@ -233,7 +240,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-willow-growing-compost", "rs-maple-growing"},
-  localized_name = {"recipe-name.rs-maple-growing-compost"}
+  localised_name = {"technology-name.rs-compost-growing", {"item-name.rs-maple-logs"}},
+  localised_description = {"technology-description.rs-compost-growing", {"tree.rs-maple-tree"}}
 })
 
 create_technology({
@@ -262,7 +270,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-maple-growing-compost", "rs-yew-growing"},
-  localized_name = {"recipe-name.rs-yew-growing-compost"}
+  localised_name = {"technology-name.rs-compost-growing", {"item-name.rs-yew-logs"}},
+  localised_description = {"technology-description.rs-compost-growing", {"tree.rs-yew-tree"}}
 })
 
 create_technology({
@@ -291,7 +300,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-yew-growing-compost", "rs-magic-growing"},
-  localized_name = {"recipe-name.rs-magic-growing-compost"}
+  localised_name = {"technology-name.rs-compost-growing", {"item-name.rs-magic-logs"}},
+  localised_description = {"technology-description.rs-compost-growing", {"tree.rs-magic-tree"}}
 })
 
 
@@ -322,7 +332,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-oak-growing-compost", "rs-supercompost"},
-  localized_name = {"recipe-name.rs-oak-growing-supercompost"}
+  localised_name = {"technology-name.rs-supercompost-growing", {"item-name.rs-oak-logs"}},
+  localised_description = {"technology-description.rs-supercompost-growing", {"tree.rs-oak-tree"}}
 })
 
 create_technology({
@@ -351,7 +362,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-oak-growing-supercompost", "rs-willow-growing-compost", "rs-supercompost"},
-  localized_name = {"recipe-name.rs-willow-growing-supercompost"}
+  localised_name = {"technology-name.rs-supercompost-growing", {"item-name.rs-willow-logs"}},
+  localised_description = {"technology-description.rs-supercompost-growing", {"tree.rs-willow-tree"}}
 })
 
 create_technology({
@@ -380,7 +392,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-willow-growing-supercompost", "rs-maple-growing-compost", "rs-supercompost"},
-  localized_name = {"recipe-name.rs-maple-growing-supercompost"}
+  localised_name = {"technology-name.rs-supercompost-growing", {"item-name.rs-maple-logs"}},
+  localised_description = {"technology-description.rs-supercompost-growing", {"tree.rs-maple-tree"}}
 })
 
 create_technology({
@@ -409,7 +422,8 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-maple-growing-supercompost", "rs-yew-growing-compost", "rs-supercompost"},
-  localized_name = {"recipe-name.rs-yew-growing-supercompost"}
+  localised_name = {"technology-name.rs-supercompost-growing", {"item-name.rs-yew-logs"}},
+  localised_description = {"technology-description.rs-supercompost-growing", {"tree.rs-yew-tree"}}
 })
 
 create_technology({
@@ -438,5 +452,6 @@ create_technology({
   },
   icon_size = 256,
   extra_prerequisites = {"rs-yew-growing-supercompost", "rs-magic-growing-compost", "rs-supercompost"},
-  localized_name = {"recipe-name.rs-magic-growing-supercompost"}
+  localised_name = {"technology-name.rs-supercompost-growing", {"item-name.rs-magic-logs"}},
+  localised_description = {"technology-description.rs-supercompost-growing", {"tree.rs-magic-tree"}}
 })