+++ mods/Farming/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [0.1.0] - 2019-07-30 - LeonSkills
 ### Added
+- Added herb farming and cleaning
+- Added herb farming technologies
+
+## [0.1.0] - 2019-07-30 - LeonSkills
+### Added
 - Added oak, willow, maple, yew and magic seeds, roots and patches
 - Within the patches you can grow logs (Requires RsWoodcutting) and roots
 - Added compost and compost bins. Some items can now be turned into compost, those items act as "compost fuel". Compost and supercompost increase yield on patches.